% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MazamaCoreUtils-package.R
\docType{data}
\name{APIKeys}
\alias{APIKeys}
\title{API keys for data services}
\format{
A named list of character strings.
}
\description{
Internal session state used to store API keys for web services.
}
\details{
Users can set API keys with \code{\link[=setAPIKey]{setAPIKey()}}. Keys are remembered for the
duration of the R session and can be retrieved with \code{\link[=getAPIKey]{getAPIKey()}}.

This provides a small abstraction layer for dependent packages so that data
access functions can test for and retrieve provider-specific API keys with
generic code.
}
\seealso{
\code{\link[=getAPIKey]{getAPIKey()}}, \code{\link[=setAPIKey]{setAPIKey()}}, \code{\link[=showAPIKeys]{showAPIKeys()}}
}
\keyword{environment}
