% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devtools_check.R
\name{check}
\alias{check}
\title{Package check}
\usage{
check(pkg = ".")
}
\arguments{
\item{pkg}{passed to \code{devtools::check()}}
}
\value{
No return.
}
\description{
Runs devtools::check() with no additional arguments.
This is a wrapper for:

\preformatted{
devtools::check()
}
}
