% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devtools_check.R
\name{check_fast}
\alias{check_fast}
\title{Check without building vignettes}
\usage{
check_fast(pkg = ".")
}
\arguments{
\item{pkg}{description, can be path or package name. See as.package() for
more information}
}
\value{
No return.
}
\description{
Runs devtools::check() with appropriate arguments to avoid
building or checking vignettes.
}
