% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devtools_check.R
\name{check_fastest}
\alias{check_fastest}
\title{Check without building vignettes or examples}
\usage{
check_fastest(pkg = ".")
}
\arguments{
\item{pkg}{description, can be path or package name. See as.package() for
more information}
}
\value{
No return.
}
\description{
Runs devtools::check() with appropriate arguments to avoid
documentation rebuild, vignettes, examples and tests.
}
