% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devtools_check.R
\name{check_slow}
\alias{check_slow}
\title{Check and run donttest{} examples}
\usage{
check_slow(pkg = ".")
}
\arguments{
\item{pkg}{description, can be path or package name. See as.package() for
more information}
}
\value{
No return.
}
\description{
Runs devtools::check() with appropriate arguments to run
donttest{} examples
}
