% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devtools_check.R
\name{check_slower}
\alias{check_slower}
\title{Check and run donttest{} and dontrun{} examples}
\usage{
check_slower(pkg = ".")
}
\arguments{
\item{pkg}{description, can be path or package name. See as.package() for
more information}
}
\value{
No return.
}
\description{
Runs devtools::check() with appropriate arguments to run
donttest{} and dontrun{} examples
}
