% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devtools_check.R
\name{check_slowest}
\alias{check_slowest}
\title{Check and run donttest{} and dontrun{} examples, max testing}
\usage{
check_slowest(pkg = ".")
}
\arguments{
\item{pkg}{description, can be path or package name. See as.package() for
more information}
}
\value{
No return.
}
\description{
Runs devtools::check() with appropriate arguments to run
donttest{} and dontrun{} examples, maximum testing with things like
--use-valgrind
}
