% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MazamaCoreUtils-package.R
\name{getAPIKey}
\alias{getAPIKey}
\title{Get API key}
\usage{
getAPIKey(provider = NULL)
}
\arguments{
\item{provider}{Web service provider.}
}
\value{
API key string, \code{NULL}, or a named list of all provider/key pairs.
}
\description{
Return the API key associated with a web service provider.
}
\details{
If \code{provider = NULL}, all currently stored API keys are returned.
}
\examples{
setAPIKey("exampleProvider", "abcd1234efgh5678")

getAPIKey("exampleProvider")
getAPIKey()

# Remove the example key
setAPIKey("exampleProvider", NULL)

}
\seealso{
\link{APIKeys}, \code{\link[=setAPIKey]{setAPIKey()}}, \code{\link[=showAPIKeys]{showAPIKeys()}}
}
\keyword{environment}
