% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-logging.R
\name{logger.info}
\alias{logger.info}
\title{Python-style logging statements}
\usage{
logger.info(msg, ...)
}
\arguments{
\item{msg}{Message with optional format strings.}

\item{...}{Additional arguments passed to \code{sprintf()} formatting.}
}
\value{
No return value. Called for side effects.
}
\description{
Emit an \code{INFO} level log message.
}
\details{
Logging must first be initialized with \code{\link[=logger.setup]{logger.setup()}}.
}
\examples{
logger.setup(infoLog = file.path(tempdir(), "INFO.log"))
logger.info("Processed \%d of \%d records", 5L, 10L)

}
\seealso{
\code{\link[=logger.setup]{logger.setup()}}
}
