% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-logging.R
\name{logger.isInitialized}
\alias{logger.isInitialized}
\title{Check whether logging has been initialized}
\usage{
logger.isInitialized()
}
\value{
Logical scalar indicating whether logging has been initialized.
}
\description{
Determine whether \code{\link[=logger.setup]{logger.setup()}} has already been called.
}
\details{
This function is useful in package code that conditionally emits log
statements only when logging has been configured.
}
\examples{
\dontrun{
logger.isInitialized()

logger.setup()

logger.isInitialized()
}

}
\seealso{
\code{\link[=logger.setup]{logger.setup()}}
}
