% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-logging.R
\docType{data}
\name{logLevels}
\alias{logLevels}
\alias{FATAL}
\alias{ERROR}
\alias{WARN}
\alias{INFO}
\alias{DEBUG}
\alias{TRACE}
\title{Log levels}
\format{
An object of class \code{integer} of length 1.
}
\usage{
FATAL
}
\description{
Logging level constants used by the MazamaCoreUtils logging API.
}
\details{
Available log levels include:

\preformatted{
FATAL
ERROR
WARN
INFO
DEBUG
TRACE
}

These constants are retained for backwards compatibility with the original
MazamaCoreUtils logging system.
}
\examples{
# Level constants are named integers
c(FATAL, ERROR, WARN, INFO, DEBUG, TRACE)
names(INFO)

# Show INFO and higher messages in the console
logger.setup()
logger.setLevel(INFO)
logger.info("Console logging is now enabled for INFO")

}
\keyword{datasets}
