% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MazamaCoreUtils.R
\docType{package}
\name{MazamaCoreUtils}
\alias{MazamaCoreUtils}
\title{Utility Functions for Production R Code}
\description{
The MazamaCoreUtils package was created by MazamaScience to
regularize our work building R-based web services.

The main goal of this package is to create an internally standardized set of
functions that we can use in various systems that are being run
operationally. Areas of functionality supported by this package include:

\itemize{
\item{ python style logging }
\item{ simple error messaging }
\item{ cache management }
\item{ date parsing }
\item{ source code linting }
}
}
