% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionArgument_lintRules.R
\docType{data}
\name{timezoneLintRules}
\alias{timezoneLintRules}
\title{Timezone linting rules}
\format{
A named list of function/argument pairs.
}
\usage{
timezoneLintRules
}
\description{
Rules used by \code{\link[=lintFunctionArgs_file]{lintFunctionArgs_file()}} and \code{\link[=lintFunctionArgs_dir]{lintFunctionArgs_dir()}} to find
date/time function calls that should explicitly specify timezone arguments.
}
\details{
Each list name is a function to check. Each value is the required named
timezone-related argument for that function.

Entries with \code{"DEPRECATED"} are used to flag functions that should generally
be avoided in package code because they depend on the local system clock or
timezone.
}
\examples{
str(timezoneLintRules)

}
\keyword{datasets}
