% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MazamaCoreUtils.R
\docType{data}
\name{APIKeys}
\alias{APIKeys}
\title{API keys for data services.}
\format{
List of character strings.
}
\description{
This package maintains an internal set of API keys which
users can set using \code{setAPIKey()}. These keys will be remembered for
the duration of an R session. This functionality provides an abstraction
layer in dependent packages so that data access functions can test for and
access specific API keys with generic code.
}
\seealso{
\link{getAPIKey}

\link{setAPIKey}

\link{showAPIKeys}
}
\keyword{environment}
