% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devtools_check.R
\name{check_faster}
\alias{check_faster}
\title{Package check without building vignettes or examples}
\usage{
check_faster(pkg = ".")
}
\arguments{
\item{pkg}{passed to \code{devtools::check()}}
}
\value{
No return.
}
\description{
Runs devtools::check() with appropriate arguments to avoid
building or checking vignettes or examples.
This is a wrapper for:

\preformatted{
devtools::check(
  build_args = c("--no-build-vignettes"),
  args = c("--ignore-vignettes", "--no-examples")
)
}
}
