% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devtools_check.R
\name{check_fastest}
\alias{check_fastest}
\title{Package check without building vignettes, examples or tests}
\usage{
check_fastest(pkg = ".")
}
\arguments{
\item{pkg}{passed to \code{devtools::check()}}
}
\value{
No return.
}
\description{
Runs devtools::check() with appropriate arguments to avoid
building or checking documentation, vignettes, examples and tests.
This is a wrapper for:

\preformatted{
devtools::check(
  build_args = c("--no-build-vignettes", "--no-manual"),
  args = c("--ignore-vignettes", "--no-manual",
           "--no-examples", "--no-tests")
)
}
}
