% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devtools_check.R
\name{check_slow}
\alias{check_slow}
\title{Check and run donttest examples}
\usage{
check_slow(pkg = ".")
}
\arguments{
\item{pkg}{passed to \code{devtools::check()}}
}
\value{
No return.
}
\description{
Runs devtools::check() with appropriate arguments to run
\code{donttest{...}} examples.
This is a wrapper for:

\preformatted{
devtools::check(
  args = c("--run-donttest")
)
}
}
