% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devtools_check.R
\name{check_slower}
\alias{check_slower}
\title{Check and run donttest and dontrun examples}
\usage{
check_slower(pkg = ".")
}
\arguments{
\item{pkg}{passed to \code{devtools::check()}}
}
\value{
No return.
}
\description{
Runs devtools::check() with appropriate arguments to run
\code{donttest{...}} and \code{dontrun{..}} examples.
This is a wrapper for:

\preformatted{
devtools::check(
  args = c("--run-donttest", "--run-dontrun")
)
}
}
