% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devtools_check.R
\name{check_slowest}
\alias{check_slowest}
\title{Check and run donttest{} and dontrun{} examples, max testing}
\usage{
check_slowest(pkg = ".")
}
\arguments{
\item{pkg}{passed to \code{devtools::check()}}
}
\value{
No return.
}
\description{
Runs devtools::check() with maximal testing options.
This is a wrapper for:

\preformatted{
devtools::check(
  args = c("--run-donttest", "--run-dontrun", "--use-gct")
)
}
}
