% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MazamaCoreUtils-package.R
\name{getAPIKey}
\alias{getAPIKey}
\title{Get API key}
\usage{
getAPIKey(provider = NULL)
}
\arguments{
\item{provider}{Web service provider.}
}
\value{
API key string or a list of provider:key pairs.
}
\description{
Returns the API key associated with a web service.
If \code{provider == NULL} a list is returned containing all recognized
API keys.
}
\seealso{
\link{APIKeys}

\link{setAPIKey}

\link{showAPIKeys}
}
\keyword{environment}
