% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initializeLogging.R
\name{initializeLogging}
\alias{initializeLogging}
\title{Initialize standard log files}
\usage{
initializeLogging(logDir = NULL, filePrefix = "", createDir = TRUE)
}
\arguments{
\item{logDir}{Directory in which to write log files.}

\item{filePrefix}{Character string prepended to log files.}

\item{createDir}{Logical specifying whether to create a missing \code{logDir}
or issue an error message.}
}
\description{
Convenience function that wraps common logging initialization steps.
}
