% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initializeLogging.R
\name{initializeLogging}
\alias{initializeLogging}
\title{Initialize standard log files}
\usage{
initializeLogging(logDir = NULL)
}
\arguments{
\item{logDir}{Directory in which to write log files.}
}
\description{
Convenience function that wraps logging initialization steps common to Mazama
Science web services.
}
