% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-logging.R
\name{logger.fatal}
\alias{logger.fatal}
\title{Python-style logging statements}
\usage{
logger.fatal(msg, ...)
}
\arguments{
\item{msg}{Message with optional format strings.}

\item{...}{Additional arguments passed to \code{sprintf()} formatting.}
}
\value{
No return value. Called for side effects.
}
\description{
Emit a \code{FATAL} level log message.
}
\details{
Logging must first be initialized with \code{\link[=logger.setup]{logger.setup()}}.
}
\seealso{
\code{\link[=logger.setup]{logger.setup()}}
}
