% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-logging.R
\name{logger.isInitialized}
\alias{logger.isInitialized}
\title{Check for initialization of loggers}
\usage{
logger.isInitialized()
}
\value{
\code{TRUE} if logging has already been initialized.
}
\description{
Returns \code{TRUE} if logging has been initialized. This allows packages
to emit logging statements only if logging has already been set up.
}
\examples{
\dontrun{
logger.isInitialized()
logger.setup()
logger.isInitialized()
}

}
\seealso{
\code{\link{logger.setup}}

\code{\link{initializeLogging}}
}
