% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-logging.R
\name{logger.trace}
\alias{logger.trace}
\title{Python-style logging statements}
\usage{
logger.trace(msg, ...)
}
\arguments{
\item{msg}{Message with format strings applied to additional arguments.}

\item{\dots}{Additional arguments to be formatted.}
}
\value{
No return value.
}
\description{
After initializing the level-specific log files with \code{logger.setup(...)},
this function will generate \code{TRACE} level log statements.
}
\note{
All functionality is built on top of the excellent \pkg{logger} package.
}
\seealso{
\code{\link{logger.setup}}
}
