% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cacheManagement.R
\name{manageCache}
\alias{manageCache}
\title{Manage the Size of a Cache}
\arguments{
\item{cacheDir}{location of cache directory}

\item{extensions}{vector of file extensions eligible for removal}

\item{maxCacheSize}{maximum cache size in megabytes}

\item{sortBy}{value to sort by when sorting eligible files for removal}
}
\value{
Invisibly returns the number of files removed.
}
\description{
If \code{cacheDir} takes up more than \code{maxCacheSize}
megabytes on disk, files will be removed in order of access time by
default. Only files matching \code{extensions} are eligible for removal.
Files can also be removed in order of change time with \code{sortBy='c'}  
or modify time with \code{sortBy='m'}. This may be useful for removing 
out of date data from a cache.
}
\examples{
# Create a cache directory and fill it with 1.6 MB of data
CACHE_DIR <- tempdir()
write.csv(matrix(1,400,500), file=file.path(CACHE_DIR,'m1.csv'))
write.csv(matrix(2,400,500), file=file.path(CACHE_DIR,'m2.csv'))
write.csv(matrix(3,400,500), file=file.path(CACHE_DIR,'m3.csv'))
write.csv(matrix(4,400,500), file=file.path(CACHE_DIR,'m4.csv'))
for (file in list.files(CACHE_DIR, full.names=TRUE)) {
  print(file.info(file)[,c(1,6)])
}
# Remove files based on access time until we get under 1 MB
manageCache(CACHE_DIR, extensions='csv', maxCacheSize=1, sortBy='a')
for (file in list.files(CACHE_DIR, full.names=TRUE)) {
  print(file.info(file)[,c(1,6)])
}
# Or remove files based on modify time
manageCache(CACHE_DIR, extensions='csv', maxCacheSize=1, sortBy='m')
for (file in list.files(CACHE_DIR, full.names=TRUE)) {
  print(file.info(file)[,c(1,6)])
}
}
