% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MazamaCoreUtils-package.R
\docType{package}
\name{MazamaCoreUtils-package}
\alias{MazamaCoreUtils}
\alias{MazamaCoreUtils-package}
\title{MazamaCoreUtils}
\description{
Utility functions for production R code.
}
\details{
MazamaCoreUtils provides a standardized set of utility functions used by
MazamaScience packages and R-based operational systems.

Functionality includes:

\itemize{
\item Python-style logging
\item simple error messaging
\item cache management
\item API key handling
\item date parsing and formatting
\item longitude/latitude validation
\item unique location ID creation
\item source code linting
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/MazamaScience/MazamaCoreUtils}
  \item Report bugs at \url{https://github.com/MazamaScience/MazamaCoreUtils/issues}
}

}
\author{
\strong{Maintainer}: Jonathan Callahan \email{jonathan.s.callahan@gmail.com}

Other contributors:
\itemize{
  \item Eli Grosman \email{eligrosman1@gmail.com} [contributor]
  \item Spencer Pease \email{spencerpease618@gmail.com} [contributor]
  \item Thomas Bergamaschi \email{tbergama@gmail.com} [contributor]
}

}
\keyword{internal}
