% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MazamaCoreUtils-package.R
\name{setAPIKey}
\alias{setAPIKey}
\title{Set API key}
\usage{
setAPIKey(provider = NULL, key = NULL)
}
\arguments{
\item{provider}{Web service provider.}

\item{key}{API key.}
}
\value{
Invisibly returns the previous value of the API key.
}
\description{
Set the API key associated with a web service provider.
}
\details{
API keys are stored in package session state and are remembered only for the
duration of the current R session.
}
\seealso{
\code{\link[=getAPIKey]{getAPIKey()}}, \code{\link[=showAPIKeys]{showAPIKeys()}}
}
\keyword{environment}
