% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MazamaCoreUtils.R
\name{setAPIKey}
\alias{setAPIKey}
\title{Set APIKey}
\usage{
setAPIKey(provider = NULL, key = NULL)
}
\arguments{
\item{provider}{Web service provider.}

\item{key}{API key.}
}
\value{
Silently returns previous value of the API key.
}
\description{
Sets the API key associated with a web service.
}
\seealso{
\link{getAPIKey}

\link{showAPIKeys}
}
\keyword{environment}
