% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionArgument_lintRules.R
\docType{data}
\name{timezoneLintRules}
\alias{timezoneLintRules}
\title{Rules for timezone linting.}
\format{
A list of function = argument pairs.
}
\usage{
timezoneLintRules
}
\description{
This set of rules is for use with the
\code{lintFunctionArgs_~()}. functions. It includes all time-related
functions from the \pkg{base} and \pkg{lubridate} packages that are involved
with parsing or formatting datetimes and helps check whether the appropriate
timezone arguments are being explicitly used.
}
\keyword{datasets}
