% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{timezoneLintRules}
\alias{timezoneLintRules}
\title{Timezone linting rules for use with function argument linting}
\format{A list where the name of an element is a function name, and the
  value of the element is the name of the argument which accepts a timezone.}
\usage{
timezoneLintRules
}
\description{
A list mapping functions accepting timezone arguments to the
  name of the argument accepting the timezone.
}
\keyword{datasets}
