% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-location.R
\name{validateLonsLats}
\alias{validateLonsLats}
\title{Validate longitude and latitude vectors}
\usage{
validateLonsLats(longitude = NULL, latitude = NULL, na.rm = FALSE)
}
\arguments{
\item{longitude}{Vector of longitudes in decimal degrees E.}

\item{latitude}{Vector of latitudes in decimal degrees N.}

\item{na.rm}{Logical specifying whether to remove \code{NA} values before
validation.}
}
\value{
Invisibly returns \code{TRUE} if no error message has been generated.
}
\description{
Longitude and latitude vectors validated to be parseable as numeric
and within the bounds -180:180 and -90:90. If validation fails, an error is
generated.
}
